+++ src/main/java/com/microsoft/sqlserver/jdbc/SQLServerConnection.java
@@ -2016,7 +2016,6 @@ public class SQLServerConnection implements ISQLServerConnection, java.io.Serial
                         throw e;
                     } else {
                         // only retry if transient error
-
                         SQLServerError sqlServerError = e.getSQLServerError();
                         if (!TransientError.isTransientError(sqlServerError)) {
                             throw e;
